pyproject.toml
[tool.poetry]
name = "flask-app-gunicorn"
version = "0.1.0"
description = ""
authors = ["JamieCattoCode <jamietcatto@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Flask = "^3.0.2"
gunicorn = "^21.2.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
